Allow negative numbers for cantor pairing
Cantor pairing only works for natural numbers. `TileMap` allows cells to have negative indices. This can actually break the existing logic. In the cantor pairing function we need to translate negative numbers to natural numbers. I chose to use similar overflow behavior you would find when wrapping an unsigned integer. When a number comes in that is negative like `-1` it gets translated to `9223372030412324865` and decreases from there.
I had to switch around order of operations in the math to divide earlier and prevent integer overflow. I spent a while debugging an issue related to this so hopefully this saves someone some effort.